chore: refactor package
